1.optimized the skia compilation process. 2.optimized the node compilation process. 3.moved Environment from vgg_application to vgg_adapter. 4.added SDL2 dll to install. 5.supported compile target viewer.
