+++ cmake/ImportSkia.cmake
@@ -73,9 +73,12 @@ else()
   set(SKIA_EXTERNAL_PROJECT_DIR ${SKIA_SOURCE_DIR} CACHE STRING "" FORCE)
 endif()
 
-if(MSVC AND NOT DEFINED SKIA_DIR)
-  message(STATUS "Run skia git-sync-deps")
-  execute_process(COMMAND "python3" "tools/git-sync-deps" WORKING_DIRECTORY "${CMAKE_SOURCE_DIR}/lib/skia")
+if(MSVC)
+  #message(STATUS "Run skia git-sync-deps")
+  #execute_process(COMMAND "python3" "tools/git-sync-deps" WORKING_DIRECTORY "${CMAKE_SOURCE_DIR}/lib/skia")
+  if(NOT EXISTS "${SKIA_SOURCE_DIR}/bin/gn.exe")
+    execute_process(COMMAND "python3" "bin/fetch-gn" WORKING_DIRECTORY ${SKIA_SOURCE_DIR} COMMAND_ERROR_IS_FATAL ANY)
+  endif()
 endif()
 
 message(STATUS "Use skia ${SKIA_EXTERNAL_PROJECT_DIR}")
