+++ src/Adapter/PlatformAdapter/Native/Environment.cpp
